Split github.go into focused files by responsibility
Break 588-line github.go into 4 files:
- github.go: type definitions and Handler methods (137 lines)
- github_install.go: Clone, Install, InstallFromRelease, Uninstall
- github_check.go: Check, checkUpdates
- gh_extension.go: GHExtension methods and manifest handling
